+++ tests/testthat/test_migration.R
@@ -170,6 +170,18 @@ test_that("migration can handle multiple expected first arrays", {
 })
 # n
 
+
+test_that("inverted spatial does not cause trouble", {
+	# Introduced when issue #192 was revealed and corrected
+	write.csv(example.spatial[nrow(example.spatial):1, ],
+	          "spatial.csv", row.names = FALSE)
+	output <- suppressWarnings(migration(tz = 'Europe/Copenhagen', GUI = "never"))
+	expect_true(TRUE) # Dummy test just so it is not considered skipped
+})
+
+write.csv(example.spatial, "spatial.csv", row.names = FALSE)
+
+
 # Throw in a fake results object just to test the number appending code
 a = 1
 save(a, file = "actel_migration_results.RData")