+++ tests/testthat/test_residency.R
@@ -134,6 +134,18 @@ test_that("residency can handle multi-sensor data", {
 # reset bio
 write.csv(example.biometrics, "biometrics.csv", row.names = FALSE)
 
+
+test_that("inverted spatial does not cause trouble", {
+	# Introduced when issue #192 was revealed and corrected
+	write.csv(example.spatial[nrow(example.spatial):1, ],
+	          "spatial.csv", row.names = FALSE)
+	output <- suppressWarnings(residency(tz = 'Europe/Copenhagen', GUI = "never"))
+	expect_true(TRUE) # Dummy test just so it is not considered skipped
+})
+
+write.csv(example.spatial, "spatial.csv", row.names = FALSE)
+
+
 # Throw in a fake results object just to test the number appending code
 a = 1
 save(a, file = "actel_residency_results.RData")